refactor the stories
