+++ src/table/stories/index.ts
@@ -1,0 +1,7 @@
+export * from "./app-custom-table.component";
+export * from "./app-expansion-table.component";
+export * from "./app-overflow-table.component";
+export * from "./app-pagination-table.component";
+export * from  "./app-skeleton-table.component";
+export * from "./app-table.component";
+export * from "./app-no-data.component";
